[u/f] 1. fixed libc stdio : ftello depended by ftell needs feature test macros       2. dev : ahci (hba) start code and port identifications       3. dev : blkdev's buffer_t - new func family `bdma_` to handle          anonymous buffers.       4. dev : pci - other strings of class names added

+++ src/app/libc/stdlib/strtoull.c
@@ -1,5 +1,5 @@
 #include <limits.h>
-typedef unsigned long TYPE;
+typedef unsigned long long TYPE;
 #define NAME strtoull
 #define MAX ULLONG_MAX
 #define MIN 0
